mc_disco_diffusion_gan/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=1.2.0
numpy>=1.24.0
scikit-image>=0.21.0
h5py>=3.9.0
PyYAML>=6.0.1
matplotlib>=3.7.0
tqdm>=4.65.0
einops>=0.7.0
